ai-engine/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-docx>=1.1.2
google-generativeai>=0.8.0
pydantic>=2.7.1
httpx>=0.27.0
python-jose>=3.3.0
python-dotenv>=1.0.1
